test(matcher): one home per fact in the withdrawal suite
`test_withdrawal_matching.py` and the per-step files were written from different angles and
met in the middle: three of its four tests are strict subsets of a step file's version,
which adds a decoy and asserts more. Its two facts that had no other home move to the step
they belong to — "an open withdrawal stays open" to the tezos-withdrawals file, "attaching
consumes both hashes" to the attachment file — and the file goes.

The dead `updated_at` assignment (auto_now overwrites it; #42) was pinned four times over.
One pin states it; the other three asserted it in passing and would have to be edited
together the day it is fixed.
