--- perf/snapshots/bench-all-patches.json
@@ -1,120 +1,0 @@
-{
- "label": "all-patches",
- "scale": 1.0,
- "passes": 3,
- "seed_seconds": 11.9,
- "requested_pools": {
-  "tezos_withdrawals": 8886,
-  "xtz_deposits": 2656,
-  "claimed_fast": 950,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16,
-  "settled_bridge_withdrawals": 26000,
-  "settled_bridge_deposits": 26000
- },
- "pools": {
-  "tezos_withdrawals": 8886,
-  "claimed_fast": 950,
-  "xtz_deposits": 2656,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16
- },
- "pools_after": {
-  "tezos_withdrawals": 8886,
-  "claimed_fast": 950,
-  "xtz_deposits": 2656,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16
- },
- "fixed_point": true,
- "steps": {
-  "tezos_deposits": {
-   "queries": 1,
-   "ms": 27.8,
-   "ms_all": [
-    30.6,
-    27.8,
-    27.4
-   ]
-  },
-  "inbox": {
-   "queries": 1,
-   "ms": 2.5,
-   "ms_all": [
-    2.8,
-    2.5,
-    2.1
-   ]
-  },
-  "michelson_deposits": {
-   "queries": 1,
-   "ms": 1.2,
-   "ms_all": [
-    1.2,
-    1.2,
-    1.0
-   ]
-  },
-  "etherlink_deposits": {
-   "queries": 4,
-   "ms": 33.5,
-   "ms_all": [
-    33.5,
-    36.2,
-    32.5
-   ]
-  },
-  "etherlink_xtz_deposits": {
-   "queries": 7,
-   "ms": 122.9,
-   "ms_all": [
-    135.1,
-    122.9,
-    115.9
-   ]
-  },
-  "etherlink_withdrawals": {
-   "queries": 1,
-   "ms": 28.1,
-   "ms_all": [
-    30.9,
-    28.1,
-    27.7
-   ]
-  },
-  "outbox": {
-   "queries": 1,
-   "ms": 2.5,
-   "ms_all": [
-    3.2,
-    2.5,
-    2.3
-   ]
-  },
-  "tezos_withdrawals": {
-   "queries": 2,
-   "ms": 204.5,
-   "ms_all": [
-    204.5,
-    213.4,
-    171.9
-   ]
-  },
-  "claimed_fast_withdrawals": {
-   "queries": 3,
-   "ms": 59.6,
-   "ms_all": [
-    60.1,
-    59.6,
-    59.1
-   ]
-  }
- },
- "total": {
-  "queries": 21,
-  "ms": 482.6
- }
-}
