--- perf/snapshots/bench-baseline.json
@@ -1,120 +1,0 @@
-{
- "label": "baseline",
- "scale": 1.0,
- "passes": 3,
- "seed_seconds": 11.6,
- "requested_pools": {
-  "tezos_withdrawals": 8886,
-  "xtz_deposits": 2656,
-  "claimed_fast": 950,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16,
-  "settled_bridge_withdrawals": 26000,
-  "settled_bridge_deposits": 26000
- },
- "pools": {
-  "tezos_withdrawals": 8886,
-  "claimed_fast": 950,
-  "xtz_deposits": 2656,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16
- },
- "pools_after": {
-  "tezos_withdrawals": 8886,
-  "claimed_fast": 950,
-  "xtz_deposits": 2656,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16
- },
- "fixed_point": true,
- "steps": {
-  "tezos_deposits": {
-   "queries": 1,
-   "ms": 8.5,
-   "ms_all": [
-    29.0,
-    7.5,
-    8.5
-   ]
-  },
-  "inbox": {
-   "queries": 1,
-   "ms": 2.4,
-   "ms_all": [
-    2.6,
-    2.1,
-    2.4
-   ]
-  },
-  "michelson_deposits": {
-   "queries": 1,
-   "ms": 1.2,
-   "ms_all": [
-    1.1,
-    1.2,
-    1.4
-   ]
-  },
-  "etherlink_deposits": {
-   "queries": 455,
-   "ms": 343.8,
-   "ms_all": [
-    357.5,
-    335.7,
-    343.8
-   ]
-  },
-  "etherlink_xtz_deposits": {
-   "queries": 2660,
-   "ms": 2368.3,
-   "ms_all": [
-    2394.5,
-    2368.3,
-    2340.9
-   ]
-  },
-  "etherlink_withdrawals": {
-   "queries": 1,
-   "ms": 8.4,
-   "ms_all": [
-    29.9,
-    7.3,
-    8.4
-   ]
-  },
-  "outbox": {
-   "queries": 1,
-   "ms": 2.1,
-   "ms_all": [
-    2.2,
-    1.8,
-    2.1
-   ]
-  },
-  "tezos_withdrawals": {
-   "queries": 8888,
-   "ms": 14153.1,
-   "ms_all": [
-    14153.1,
-    14008.0,
-    14885.5
-   ]
-  },
-  "claimed_fast_withdrawals": {
-   "queries": 952,
-   "ms": 601.0,
-   "ms_all": [
-    587.5,
-    601.0,
-    603.0
-   ]
-  }
- },
- "total": {
-  "queries": 12960,
-  "ms": 17488.8
- }
-}
