--- perf/snapshots/bench-p1-tezos-withdrawals.json
@@ -1,120 +1,0 @@
-{
- "label": "p1-tezos-withdrawals",
- "scale": 1.0,
- "passes": 3,
- "seed_seconds": 11.6,
- "requested_pools": {
-  "tezos_withdrawals": 8886,
-  "xtz_deposits": 2656,
-  "claimed_fast": 950,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16,
-  "settled_bridge_withdrawals": 26000,
-  "settled_bridge_deposits": 26000
- },
- "pools": {
-  "tezos_withdrawals": 8886,
-  "claimed_fast": 950,
-  "xtz_deposits": 2656,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16
- },
- "pools_after": {
-  "tezos_withdrawals": 8886,
-  "claimed_fast": 950,
-  "xtz_deposits": 2656,
-  "etherlink_deposits": 453,
-  "open_bridge_withdrawals": 26,
-  "open_bridge_deposits": 16
- },
- "fixed_point": true,
- "steps": {
-  "tezos_deposits": {
-   "queries": 1,
-   "ms": 36.7,
-   "ms_all": [
-    36.7,
-    28.1,
-    52.9
-   ]
-  },
-  "inbox": {
-   "queries": 1,
-   "ms": 2.6,
-   "ms_all": [
-    2.8,
-    2.1,
-    2.6
-   ]
-  },
-  "michelson_deposits": {
-   "queries": 1,
-   "ms": 1.2,
-   "ms_all": [
-    1.3,
-    1.0,
-    1.2
-   ]
-  },
-  "etherlink_deposits": {
-   "queries": 455,
-   "ms": 382.4,
-   "ms_all": [
-    373.2,
-    384.9,
-    382.4
-   ]
-  },
-  "etherlink_xtz_deposits": {
-   "queries": 2660,
-   "ms": 2581.6,
-   "ms_all": [
-    2606.2,
-    2545.8,
-    2581.6
-   ]
-  },
-  "etherlink_withdrawals": {
-   "queries": 1,
-   "ms": 26.1,
-   "ms_all": [
-    28.9,
-    26.1,
-    25.0
-   ]
-  },
-  "outbox": {
-   "queries": 1,
-   "ms": 2.4,
-   "ms_all": [
-    2.4,
-    2.2,
-    2.6
-   ]
-  },
-  "tezos_withdrawals": {
-   "queries": 2,
-   "ms": 169.5,
-   "ms_all": [
-    205.2,
-    166.7,
-    169.5
-   ]
-  },
-  "claimed_fast_withdrawals": {
-   "queries": 952,
-   "ms": 660.5,
-   "ms_all": [
-    660.5,
-    718.9,
-    649.5
-   ]
-  }
- },
- "total": {
-  "queries": 4074,
-  "ms": 3863.0
- }
-}
